requirements.txt
backpack-for-pytorch==1.3.0
bayesian-optimization==1.2.0
black==21.11b1
click==8.0.3
cycler==0.11.0
einops==0.3.2
fonttools==4.28.1
gym==0.19.0
importlib-metadata==4.8.2
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.5.2
mujoco-py==2.1.2.14
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.2
palettable==3.3.0
pandas==1.3.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
threadpoolctl==3.0.0
tikzplotlib==0.9.15
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
typed-ast==1.5.0
typing_extensions==4.0.0
zipp==3.6.0
